process-mate/backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
python-chess>=1.9.0
gunicorn>=20.1.0